requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
Jinja2==3.1.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4 # For password hashing
python-jose==3.3.0 # For JWT, if needed later (even if sessions are primary)
python-dotenv==1.0.1
aiofiles==23.2.1 # For serving static files
websockets==12.0 # for potential live updates
pydantic-settings==2.2.1 # for managing settings
httpx==0.27.0 # for testing, can be moved to dev-requirements later
# For Leaflet.js and Tailwind CSS, these are frontend dependencies, not Python.
# We'll include them via CDN or build process.
